Store version in a js file
We were previously storing the version in a JSON file. But, requiring
a JSON file causes trouble to people using Webpack to bundle the NPM
package. Because Webpack can only process JavaScript by default, they
have to figure out that json-loader, or something similar, needs to be
added in order to bundle auth0.js (or any other library that depends
on it, such as lock or lock-passwordless).