Trim `username`, `email` and `phoneNumber` params in every request (#895)
This is in reference to issue #872

It's possible there are additional places that username/email should be trimmed. If so, you can consider this a WIP 😉; but as I'm still getting used to the code base, I wanted to get this in to make sure I'm on the right track.